1. added the boxplot and comparative bar chart options to the data_visualization_basic GUI; 2. consolidated all charts in a single charts_util; 3. fixed a bug in Sankey chart when using 2 variables only.
